+++ LoopbackDecode.m
@@ -55,12 +55,12 @@ err:
     return;
 }
 
-- (unsigned int)frontSize
+- (size_t)frontSize
 {
     return sizeof(loopback_type);
 }
 
-- (unsigned int)rearSize
+- (size_t)rearSize
 {
     return 0;
 }
@@ -140,7 +140,7 @@ err:
     return YES;
 }
 
-- (unsigned int)numberOfChildren
+- (size_t)numberOfChildren
 {
     return 1;
 }
@@ -160,7 +160,7 @@ err:
     return [ret autorelease];
 }
 
-- (unsigned int)numberOfValues
+- (size_t)numberOfValues
 {
     return 1;
 }
